refactor(reflection): use canonical module layout
