+++ reflexio/server/services/reflection/components/__init__.py
@@ -1,0 +1,7 @@
+"""Reflection service internal components."""
+
+from reflexio.server.services.reflection.components.extractor import (
+    ReflectionExtractor,
+)
+
+__all__ = ["ReflectionExtractor"]
