+++ tests/e2e_tests/test_failure_path_polarity_e2e.py
@@ -47,7 +47,7 @@ from reflexio.models.api_schema.domain.entities import (
 from reflexio.models.api_schema.domain.enums import Status
 from reflexio.models.config_schema import Config, ReflectionConfig
 from reflexio.server.api_endpoints.request_context import RequestContext
-from reflexio.server.services.reflection.reflection_service import ReflectionService
+from reflexio.server.services.reflection.service import ReflectionService
 from reflexio.server.services.reflection.reflection_service_utils import (
     ReflectionDecision,
     ReflectionOutput,
