+++ tests/server/services/reflection/test_reflection_profile_supersede_integration.py
@@ -15,12 +15,12 @@ import pytest
 from reflexio.models.api_schema.domain.entities import Citation, UserProfile
 from reflexio.models.api_schema.domain.enums import ProfileTimeToLive, Status
 from reflexio.server.api_endpoints.request_context import RequestContext
-from reflexio.server.services.reflection.reflection_service import ReflectionService
 from reflexio.server.services.reflection.reflection_service_utils import (
     ReflectionDecision,
     ReflectionOutput,
     ReflectionServiceRequest,
 )
+from reflexio.server.services.reflection.service import ReflectionService
 
 pytestmark = pytest.mark.integration
